+++ src/components/Members/EditMembers.tsx
@@ -1,14 +1,11 @@
-import { FC, useCallback, useEffect, useMemo, useState } from 'react';
-import { Button, Dialog, FlexBox, Input, Label } from '@ui5/webcomponents-react';
+import { FC, useCallback, useState } from 'react';
+import { Button, FlexBox } from '@ui5/webcomponents-react';
 import { MemberTable } from './MemberTable.tsx';
-import { Member, MemberRoles, memberRolesOptions } from '../../lib/api/types/shared/members';
+import { Member } from '../../lib/api/types/shared/members';
 import { useTranslation } from 'react-i18next';
 import styles from './Members.module.css';
-import { RadioButtonsSelect, RadioButtonsSelectOption } from '../Ui/RadioButtonsSelect/RadioButtonsSelect.tsx';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
-import FadeVisibility from '../Ui/FadeVisibility/FadeVisibility.tsx';
+import { RadioButtonsSelectOption } from '../Ui/RadioButtonsSelect/RadioButtonsSelect.tsx';
+import { AddEditMemberDialog } from './AddEditMemberDialog.tsx';
 
 export interface EditMembersProps {
   members: Member[];
@@ -22,191 +19,7 @@ export const ACCOUNT_TYPES: RadioButtonsSelectOption[] = [
   { value: 'ServiceAccount', label: 'Service Account', icon: 'machine' },
 ];
 
-interface AddEditMemberDialogProps {
-  open: boolean;
-  onClose: () => void;
-  onSave: (member: Member, isEdit: boolean) => void;
-  existingMembers: Member[];
-  memberToEdit?: Member;
-}
-
 export type AccountType = 'User' | 'ServiceAccount';
-
-type MemberFormData = {
-  accountType: AccountType;
-  name: string;
-  role: string;
-  namespace?: string;
-};
-
-const AddEditMemberDialog: FC<AddEditMemberDialogProps> = ({
-  open,
-  onClose,
-  onSave,
-  existingMembers,
-  memberToEdit,
-}) => {
-  const { t } = useTranslation();
-  const isEdit = !!memberToEdit;
-
-  const memberFormSchema = useMemo(
-    () =>
-      z
-        .object({
-          accountType: z.enum(['User', 'ServiceAccount']),
-          name: z.string(),
-          role: z.string(),
-          namespace: z.string().optional(),
-        })
-        .superRefine((data, ctx) => {
-          const trimmed = data.name.trim();
-          if (!trimmed) {
-            ctx.addIssue({
-              code: z.ZodIssueCode.custom,
-              path: ['name'],
-              message: t('validationErrors.required'),
-            });
-          }
-          if (existingMembers.some((m) => m.name === trimmed && (!memberToEdit || trimmed !== memberToEdit.name))) {
-            ctx.addIssue({
-              code: z.ZodIssueCode.custom,
-              path: ['name'],
-              message: t('validationErrors.userExists'),
-            });
-          }
-        }),
-    [t, existingMembers, memberToEdit],
-  );
-
-  const {
-    register,
-    handleSubmit,
-    watch,
-    setValue,
-    reset,
-    formState: { errors },
-  } = useForm<MemberFormData>({
-    resolver: zodResolver(memberFormSchema),
-    mode: 'onChange',
-    defaultValues: {
-      accountType: 'User',
-      name: '',
-      role: MemberRoles.view,
-      namespace: '',
-    },
-  });
-
-  const accountType = watch('accountType');
-  const role = watch('role');
-
-  useEffect(() => {
-    if (open) {
-      if (memberToEdit) {
-        reset({
-          name: memberToEdit.name,
-          role: memberToEdit.role || MemberRoles.view,
-          accountType: memberToEdit.kind === 'User' ? 'User' : 'ServiceAccount',
-          namespace: memberToEdit?.namespace || '',
-        });
-      } else {
-        reset({
-          accountType: 'User',
-          name: '',
-          role: MemberRoles.view,
-          namespace: '',
-        });
-      }
-    }
-  }, [open, memberToEdit, reset]);
-
-  const onFormSubmit = (data: MemberFormData) => {
-    const trimmedName = data.name.trim();
-
-    const newMember: Member = {
-      name: trimmedName,
-      role: data.role,
-      kind: data.accountType,
-      ...(data.accountType === 'ServiceAccount' && data.namespace && { namespace: data.namespace }),
-    };
-
-    onSave(newMember, isEdit);
-    onClose();
-  };
-
-  const dialogHeader = memberToEdit ? t('EditMembers.editHeader') : t('EditMembers.addHeader') || 'Add member';
-
-  return (
-    <Dialog open={open} headerText={dialogHeader}>
-      <div className={styles.container}>
-        <FlexBox alignItems={'Baseline'} direction={'Column'} className={styles.wrapper}>
-          <FlexBox alignItems={'Baseline'} justifyContent={'SpaceBetween'}>
-            <RadioButtonsSelect
-              label={'Account type:'}
-              selectedValue={accountType}
-              options={ACCOUNT_TYPES}
-              handleOnClick={(value) => setValue('accountType', value as AccountType, { shouldValidate: true })}
-            />
-          </FlexBox>
-        </FlexBox>
-        <FlexBox direction="Column" alignItems="Stretch" className={styles.wrapper}>
-          <Label for="member-email-input">{t('common.name')}</Label>
-          <Input
-            className={styles.input}
-            id="member-email-input"
-            type={accountType === 'User' ? 'Email' : 'Text'}
-            {...register('name')}
-            valueState={errors.name ? 'Negative' : 'None'}
-            valueStateMessage={<span>{errors.name?.message}</span>}
-            data-testid="member-email-input"
-          />
-        </FlexBox>
-        <FlexBox alignItems="Stretch" direction={'Column'}>
-          <div className={styles.wrapper}>
-            <RadioButtonsSelect
-              selectedValue={role}
-              options={memberRolesOptions}
-              handleOnClick={(value) => setValue('role', value, { shouldValidate: true })}
-              label={t('MemberTable.columnRoleHeader')}
-            />
-          </div>
-
-          <div className={styles.placeholder}>
-            <FadeVisibility show={accountType === 'ServiceAccount'}>
-              <div>
-                <FlexBox direction="Column">
-                  <Label for="namespace-input">{t('common.namespace')}</Label>
-                  <Input
-                    type="Text"
-                    {...register('namespace')}
-                    className={styles.input}
-                    data-testid="namespace-input"
-                    id="namespace-input"
-                  />
-                </FlexBox>
-              </div>
-              <Label>{t('EditMembers.defaultNamespaceInfo')}</Label>
-            </FadeVisibility>
-          </div>
-
-          <Button className={styles.wrapper} onClick={onClose}>
-            {t('buttons.cancel')}
-          </Button>
-          <Button
-            className={styles.addButton}
-            data-testid="add-member-button"
-            design={'Emphasized'}
-            icon={'sap-icon://add-employee'}
-            onClick={() => {
-              handleSubmit(onFormSubmit)();
-            }}
-          >
-            {memberToEdit ? t('EditMembers.saveButton') : t('EditMembers.addButton')}
-          </Button>
-        </FlexBox>
-      </div>
-    </Dialog>
-  );
-};
 
 export const EditMembers: FC<EditMembersProps> = ({
   members,
